+++ android/app/src/main/kotlin/com/mankou/custom_input_formatter/CustomInputFormatterPlugin.kt
@@ -1,5 +1,0 @@
-package com.mankou.custom_input_formatter.custom_input_formatter
-
-import io.flutter.embedding.android.FlutterActivity
-
-class CustomInputFormatterPlugin: FlutterActivity()